+++ src/activities/GenerateQueryWhereClause.ts
@@ -1,16 +1,8 @@
 import type { IActivityHandler } from "@vertigis/workflow";
 import { DisplayFormOutputs } from "@vertigis/workflow/activities/forms/DisplayForm";
-import {
-    isDateRangeRef,
-    isDateTimeRef,
-    isFilesRef,
-    isGeometryRef,
-    isItemsRef,
-    isNumberRef,
-    isScanRef,
-} from "@vertigis/workflow/forms/utils";
 import { defs } from "@vertigis/workflow/forms/FormHost";
 import { EsriFieldType, SearchField as QueryField } from "./interfaces";
+import { getDefinitionExpression, getFormValue } from "./utils";
 
 export interface GenerateWhereClauseInputs {
     /**
@@ -18,9 +10,9 @@ export interface GenerateWhereClauseInputs {
      * @required
      */
     layer:
-        | __esri.FeatureLayer
-        | __esri.SubtypeGroupLayer
-        | __esri.SubtypeSublayer;
+    | __esri.FeatureLayer
+    | __esri.SubtypeGroupLayer
+    | __esri.SubtypeSublayer;
     /**
      * @description The form used to collect the query inputs.
      * @required
@@ -64,7 +56,7 @@ export default class GenerateWhereClause implements IActivityHandler {
         }
 
         let where = whereClause && whereClause.length > 0 ? whereClause : "1=1";
-        const definitionExpression = this.getDefinitionExpression(layer);
+        const definitionExpression = getDefinitionExpression(layer);
         if (definitionExpression) {
             where = `${where} AND ${definitionExpression}`;
         }
@@ -110,38 +102,13 @@ export default class GenerateWhereClause implements IActivityHandler {
         queryField: QueryField,
         field: __esri.Field,
     ): string | undefined {
-        const hasValue = formElement.value || formElement.value === 0;
-
-        if (hasValue) {
-            const currentValue =
-                formElement.type === "CheckBox"
-                    ? (Number(formElement.checked) as defs.Value)
-                    : (formElement.value as defs.Value);
-            if (isDateRangeRef(currentValue)) {
-                return this.formatDateRange(currentValue);
-            } else if (isDateTimeRef(currentValue)) {
-                return this.formatDate(currentValue);
-            } else if (formElement.type === "NumberRangeSlider") {
-                return this.formatNumberRange(currentValue);
-            } else if (isItemsRef(currentValue)) {
-                return this.formatItems(currentValue, field);
-            } else if (isNumberRef(currentValue)) {
-                return currentValue.numeric.toString();
-            } else if (
-                isGeometryRef(currentValue) ||
-                isFilesRef(currentValue) ||
-                isScanRef(currentValue)
-            ) {
-                throw new Error(
-                    `Unsupported form element value: ${queryField.field}: ${
-                        (currentValue as any).type
-                    }`,
-                );
-            }
+        const currentValue = getFormValue(formElement, queryField);
+        if (currentValue) {
             return this.formatValue(currentValue, queryField, field);
         }
         return undefined;
     }
+
 
     private appendToWhere(
         where: string,
@@ -158,7 +125,7 @@ export default class GenerateWhereClause implements IActivityHandler {
         return `${where}${fieldAndValue}`;
     }
     private formatValue(
-        value: defs.Value,
+        value: any,
         searchField: QueryField,
         field: __esri.Field,
     ): string | undefined {
@@ -167,90 +134,21 @@ export default class GenerateWhereClause implements IActivityHandler {
             case EsriFieldType.Guid:
             case EsriFieldType.GlobalId:
             case EsriFieldType.String:
-                if (searchField.operator.toUpperCase() === "LIKE") {
-                    formattedValue = `%${value as any}%`;
+                if (Array.isArray(value)) {
+                    formattedValue = value.map(x => `'${x}'`).join(",");
+                } else if (searchField.operator.toUpperCase() === "LIKE") {
+                    formattedValue = `%${value}%`;
                 } else {
-                    formattedValue = `'${value as any}'`;
+                    formattedValue = `'${value}'`;
                 }
                 break;
             default:
-                formattedValue = `${value as any}`;
+                if (Array.isArray(value)) {
+                    formattedValue = value.join(",");
+                } else {
+                    formattedValue = `${value}`;
+                }
                 break;
-        }
-        return formattedValue;
-    }
-
-    private getDefinitionExpression(
-        layer:
-            | __esri.FeatureLayer
-            | __esri.SubtypeGroupLayer
-            | __esri.SubtypeSublayer,
-    ): string | undefined {
-        let expression;
-        switch (layer.type) {
-            case "subtype-sublayer":
-                expression = layer.parent.definitionExpression
-                    ? `${layer.parent.definitionExpression} AND ${layer.parent.subtypeField} = ${layer.subtypeCode}`
-                    : `${layer.parent.subtypeField} = ${layer.subtypeCode}`;
-                break;
-            case "subtype-group":
-            case "feature":
-                expression = layer.definitionExpression;
-                break;
-        }
-
-        return expression;
-    }
-
-    private formatDateRange(value: defs.Value): string | undefined {
-        if (isDateRangeRef(value)) {
-            const values = [value.startDate, value.endDate];
-            return `DATE '${values[0].toISOString().split("T")[0]}' AND DATE '${
-                values[1].toISOString().split("T")[0]
-            }'`;
-        }
-        return undefined;
-    }
-
-    private formatDate(value: defs.Value): string | undefined {
-        if (isDateTimeRef(value)) {
-            return `TIMESTAMP '${new Date(value.value)
-                .toISOString()
-                .slice(0, 19)
-                .replace("T", " ")}'`;
-        }
-        return undefined;
-    }
-
-    private formatItems(
-        value: defs.Value,
-        field: __esri.Field,
-    ): string | undefined {
-        let formattedValue;
-        if (isItemsRef(value)) {
-            switch (field.type) {
-                case "guid":
-                case "global-id":
-                case "string":
-                    formattedValue = value.items
-                        .map((x) => `'${x.value as any}'`)
-                        .join(",");
-                    break;
-                default:
-                    formattedValue = value.items
-                        .map((x) => `${x.value as any}`)
-                        .join(",");
-                    break;
-            }
-        }
-        return formattedValue;
-    }
-
-    private formatNumberRange(value: defs.Value): string | undefined {
-        let formattedValue;
-
-        if (Array.isArray(value) && value.length === 2) {
-            formattedValue = `${value[0]} AND ${value[1]}`;
         }
         return formattedValue;
     }
